feat(server): Add network server and network client manager, processing packets.

+++ src/server/sync_queue.hpp
@@ -44,7 +44,7 @@ public:
 
     } // Lock is automatically released here
 
-    int count()
+    size_t count()
     {
         return m_queue.size();
     }
